delegate: change return type of callbacks to uintptr
The `NewCallback` function expects the received argument to be a
function with one uintptr-sized result. That uintptr type may be a 32 or
64 bit pointer depending on the system, but some functions were always
returning a uint64 value. That was causing errors in 32 bit systems.

fixes #83

+++ internal/delegate/delegate.go
@@ -113,17 +113,19 @@ func invoke(instancePtr, rawArgs0, rawArgs1, rawArgs2, rawArgs3, rawArgs4, rawAr
 	return instance.Invoke(instancePtr, rawArgs0, rawArgs1, rawArgs2, rawArgs3, rawArgs4, rawArgs5, rawArgs6, rawArgs7, rawArgs8)
 }
 
-func addRef(instancePtr unsafe.Pointer) uint64 {
+func addRef(instancePtr unsafe.Pointer) uintptr {
 	instance, ok := getInstance(instancePtr)
 	if !ok {
 		// instance not found
 		return ole.E_FAIL
 	}
 
-	return instance.AddRef()
+	// this may be cast to a 32 bit value,
+	// but the result from AddRef should always fit
+	return uintptr(instance.AddRef())
 }
 
-func release(instancePtr unsafe.Pointer) uint64 {
+func release(instancePtr unsafe.Pointer) uintptr {
 	instance, ok := getInstance(instancePtr)
 	if !ok {
 		// instance not found
@@ -135,5 +137,8 @@ func release(instancePtr unsafe.Pointer) uint64 {
 		// remove this delegate
 		removeInstance(instancePtr)
 	}
-	return rem
+
+	// this may be cast to a 32 bit value,
+	// but the result from AddRef should always fit
+	return uintptr(rem)
 }
